part4-django/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0

